+++ app/_layout.tsx
@@ -5,11 +5,22 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import dayjs from 'dayjs';
 import 'dayjs/locale/fr';
 import { useFonts } from 'expo-font';
+import * as Notifications from 'expo-notifications';
 import { Slot } from 'expo-router';
 import React, { useEffect } from 'react';
 import Toast from 'react-native-toast-message';
 
 dayjs.locale('fr');
+
+Notifications.setNotificationHandler({
+  handleNotification: async () => ({
+    shouldShowAlert: true, // legacy (Android / compat)
+    shouldShowBanner: true, // iOS foreground banner
+    shouldShowList: true, // iOS notification center
+    shouldPlaySound: true,
+    shouldSetBadge: false,
+  }),
+});
 
 const queryClient = new QueryClient({
   defaultOptions: {