+++ src/features/notification/utils/open-notification-settings.ts
@@ -1,0 +1,9 @@
+import { Linking, Platform } from 'react-native';
+
+export const openNotificationSettings = () => {
+  if (Platform.OS === 'ios') {
+    Linking.openURL('app-settings:');
+  } else {
+    Linking.openSettings();
+  }
+};